Add cts-db-run
shader-db is an interesting tool to test the impact of changes in Mesa
code that can affect the final shader assembly code.

Unfortunately, shader-db uses OpenGL, so if the changes are done in the
Vulkan part, it won't show any result.

This tool tries to emulate the same behaviour, but using the shaders in
Vulkan/OpenGL CTS testsuite.

The output is the same as the output obtained with shader-db/run, so
shader-db/report tool can be used later to generate the final report.