fps-summary: handle change on the output format
So far the script assumed that the txt file would start with the
following header:

"fps, frame timing (us)"

That would define the content of the rest of the lines, but recently
it changed to:
"device, format, fps, frame timing (us)"

With this change we don't hardcode the position were we read the fps
value, but we first search for the "fps" word on the header to know
their position.

This allows to support the old format, the new format, and hopefully
any new format.