fps-analysis: add --headless option on run-all-traces
This runs both apitrace/gfxreconstruct using their headless
option. This makes runs less dependent of the compositor and the
desktop resolution (although the latter still affects).

Note that as run-all-traces is intended mostly for performance
comparison, it would make sense to have this option enabled as
default. But that would be confusing for anyone using the script for
the first time, as by default, would not show anything on screen.

As I'm here, I also cleaned up the definition for --rebind and
--disable-cache-run options.