deqp: add script to get the regressions from the deqp-runner csv files
Last time I checked, deqp-runner didn't have a tool to compare the
outcome of two different runs. The way to get the regressions was
passing a base reference on the second run, but sometimes you already
finished the runs.

This is a little 30min script that allows you to do that.