deqp: add run-vulkan-cts-with-deqp-runner script
Runs all the must pass CTS tests for vulkan, but using deqp-runner.

As the run as vulkan cts script, it uses the
V3D_DEBUG=opt_compile_time envvar.

This is v3dv specific, but as it helps there, and doesn't hurt with
any other driver, we have it for convenience.